# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Connect Chat Apps API",
        "description": "Dokumentasi API untuk Aplikasi Real-Time Chat & Voice Connect",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Ambil Profil User yang Sedang Login",
                "operationId": "bbce5a07f65000c13d49ac97d9cd845f",
                "responses": {
                    "200": {
                        "description": "Informasi profil user yang sedang login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / Token tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Detail Profil User Berdasarkan ID atau Username",
                "operationId": "3efafdaa7126938df2b62bd706bba458",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User UUID atau @username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil user ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/profile": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update Profil & Status Pengguna",
                "operationId": "26397611ed0225d567e1d885ce3a5fc6",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fajar Muhammad"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "fajar_dev"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "Fajar"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Software Engineer & Gamer"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+628123456789"
                                    },
                                    "avatar_url": {
                                        "type": "string",
                                        "example": "https://example.com/avatar.png"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "online",
                                        "enum": [
                                            "online",
                                            "offline",
                                            "away",
                                            "busy",
                                            "dnd"
                                        ]
                                    },
                                    "custom_status": {
                                        "type": "string",
                                        "example": "Coding Laravel 🚀"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profil berhasil diperbarui."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi data gagal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/search": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Cari Pengguna Berdasarkan Nama atau Username",
                "operationId": "870e0c5e700bcc087b5ca3cfe0049939",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Kata kunci pencarian (nama, display_name, atau username)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "fajar"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah data per halaman (default: 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar user ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Parameter pencarian tidak valid / kosong"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Sanctum kamu di sini",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}